reverted to legacy params
