--- x/txfees/keeper/msg_server.go
@@ -1,32 +1,0 @@
-package keeper
-
-import (
-	"context"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	"github.com/osmosis-labs/osmosis/v15/x/txfees/types"
-)
-
-type msgServer struct {
-	Keeper
-}
-
-// NewMsgServerImpl returns an implementation of the MsgServer interface
-// for the provided Keeper.
-func NewMsgServerImpl(keeper Keeper) types.MsgServer {
-	return &msgServer{Keeper: keeper}
-}
-
-var _ types.MsgServer = msgServer{}
-
-// UpdateParams implements the Msg/UpdateParams message type.
-func (k msgServer) UpdateParams(goCtx context.Context, msg *types.MsgUpdateParams) (*types.MsgUpdateParamsResponse, error) {
-	// fixme: check for authorit
-	ctx := sdk.UnwrapSDKContext(goCtx)
-
-	// fixme: validate params
-	k.SetParams(ctx, msg.Params)
-
-	return &types.MsgUpdateParamsResponse{}, nil
-}
