+++ x/txfees/types/keys.go
@@ -20,5 +20,4 @@ const (
 var (
 	BaseDenomKey         = []byte("base_denom")
 	FeeTokensStorePrefix = []byte("fee_tokens")
-	ParamsKey            = []byte("params")
 )
